+++ assets/js/api.js
@@ -1,0 +1,1 @@
+export const API_URL = "https://6712d5086c5f5ced6624c443.mockapi.io/posts";
