+++ assets/style.css
@@ -183,7 +183,7 @@ body {
     color: var(--primary-color);
     padding: 1.8rem;
     border-radius: 0.3125rem;
-    max-width: 70vh;
+    max-width: 69vh;
     margin: 0 auto;
     box-sizing: border-box;
     line-height: 1.5;
