+++ index.html
@@ -31,7 +31,7 @@
         </section>
 
         <section class="image-grid">
-            <article>
+            <!-- <article>
                 <figure>
                     <img src="photos/1.jpeg" alt="A Torre Eiffel iluminada à noite, com milhares de luzes cintilando, criando um espetáculo mágico em Paris." />
                 </figure>
@@ -60,7 +60,7 @@
                 <figure>
                     <img src="photos/6.jpeg" alt="O Cristo Redentor, no Rio de Janeiro, com seus braços abertos, vigia a cidade e acolhe a todos, com sua imponência e significado religioso." />
                 </figure>
-            </article>
+            </article> -->
         </section>
     </main>
     <div id="modal" class="modal">
@@ -68,6 +68,7 @@
         <img class="modal-content" id="modal-img">
         <div id="caption"></div>
     </div>
-    <script src="assets/js/modal.js"></script>
+    <script src="assets/js/api.js" type="module"></script>
+    <script src="assets/js/index.js" type="module"></script>
 </body>
 </html>
